+++ src/getLicenses.js
@@ -80,7 +80,7 @@ If adding back to `LicenseInfo`
 /* eslint-disable max-len */
 /**
  * @param {PlainObject} cfg
- * @param {LicenseBadgerOptions#licenseInfoPath} [cfg.licenseInfoPath=resolve(process.cwd(), './licenseInfoPath.json')]
+ * @param {LicenseBadgerOptions#licenseInfoPath} [cfg.licenseInfoPath=resolve(process.cwd(), './licenseInfo.json')]
  * Not used, nor default obtained, when `filter` is `false`.
  * @param {LicenseBadgerOptions#packagePath} [cfg.packagePath=process.cwd()]
  * @param {LicenseBadgerOptions#corrections} [cfg.corrections=false]